NahaGraphs.jl

NahaGraphs provides some rudimentary functionality for building directed graphs and running GraphViz dot to visualize them.

The query function provides a simplemined way to search the graph's edges.

Generic Graph definitions

NahaGraphs.DiGraph — Type

DiGraph represents a directed graph.

Any Julia object can serve as the node of a DiGraph so long as it has identity.

Use add_edge! of a from node and a to node to add an edge (and those nodes if not already present) to a DiGraph.

The nodes, edges and query functions can be used to examine the DiGraph.

NahaGraphs.add_edge! — Function

add_edge!(graph, from, to)
add_edge!(::DiGraph, ::Pair)

Add an edge to the DiGraph going between the specified nodes. The graph is returned.

NahaGraphs.remove_edge! — Function

remove_edge!(graph, from, to)
remove_edge!(::DiGraph, ::Pair)

Remove the edge of the DiGraph going between the specified nodes. The graph is returned.

NahaGraphs.nodes — Function

nodes(graph)

Return a collection of all of the nodes of graph.

NahaGraphs.edges — Function

edges(graph)

Return a collection of all of the edges of graph. Each element of the collections is a Pair associating one nodeof the graph with another.

NahaGraphs.query — Function

query(::DiGraph, from, to)

Return the Set of edges that match the query. from and to can be objects or types, so, for example, query(graph, Number, :a) returns all of the edges that go from any Number to the Symbol :a.

To define your own graph type that delegates to DiGraph, see test/mygraph.jl. You might do this in order to streamline how nodes and edges are managed for your application.

Transforming Graphs

This is experimental.

See the "transformingGraph! and applyRule!" testset for an example.

NahaGraphs.applyRule! — Function

applyRule!(graph, rule)

Apply rule to each node of graph and modify the graph's edges as determined by rule. Rule is a function that takes the graph and a node as arguments and returns the set of edges to add and the set of edges to remove.

NahaGraphs.transformingGraph! — Function

transformingGraph!(f)

Calls f with three arguments:

a check function of one argument which will perform a non-local return if that argument is false;

an add function which is used to collect edges that should be added to a graph;

a remove function used to accumulate edges that should be removed from a graph.

transformingGraph! returns the accumulated add and remove edge sets for its caller to act on. These sets can be passed to transform! along with the graph to perform those modifications on the graph. See applyRule!.

NahaGraphs.NonlocalReturn — Type

NonlocalReturn is a type of exception that allows us to use throw and catch to implement nonlocal return to a more distant stack frame.

Usage:

begin
    exittag = NonlocalReturn()
    # An example recursive function which might need to return
    # to this context from across many stack frames:
    function f()
        if rand(1:5) == 1
            throw(exitrtag)
        end
        f()
        f()
    end
    try
	f()
    catch e
	if e != exittag
	    rethrow(e)
	end
    end
    # f has done a non-local return
end

NahaGraphs.transform! — Function

transform!(graph, addarcs, removearcs)

Remove the removearcs from graph then add the addarcs to graph, then

Definitions Specific to GraphViz dot

See the "rundot output file type" testset for an example of how to construct a graph and render it in SVG.

NahaGraphs.dotgraph — Function

dotgraph(path::String, graph, dotstyle)

Write the graph to the specified file. If the file extension is dot then a GraphViz dot file is written. Otherwise a dot description of graph is piped through the dot command, the output of which will be written to psth.

NahaGraphs.dotID — Function

dotID(node)

Return a string to be used as the id of node in a GraphViz dot file.

NahaGraphs.dotnode — Function

dotnode(io::IO, graph, dotstyle, node)

Write a Dot node statement to io describing node.

NahaGraphs.dotedge — Function

dotedge(io::IO, graph, dotstyle, from, to)

Write a Dot edge statement to io describing an edge from from to to.

NahaGraphs.graph_attributes — Function

graph_attributes(dotstyle)

Return a Dict of graph level attributes (for the Dot graph statement).

NahaGraphs.node_attributes — Function

node_attributes(dotstyle)

Return a Dict of the default node attributes that should apply to all nodes of graph.

node_attributes(dotstyle, node)

Return a Dict of the Dot attributes to be used when rendering node.

NahaGraphs.edge_attributes — Function

edge_attributes(dotstyle)

Return a Dict of the default attributes that should apply to all edges of graph.

edge_attributes(dotstyle, from, to)

Return a Dict of the Dot attributes to be used when rendering the edge running from from to to, which are both nodes.

NahaGraphs.dotescape — Function

dotescape(::AbstractString)::AbstractString

Escape an ID in the GraphViz Dot language. 'ID' is the fundamental token in Dot.

NahaGraphs.diarc — Function

diark is a convenience function for dotedge methods to call.